streamlit_mini/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
blinker==1.5
cachetools==5.3.0
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
diffusers==0.12.1
entrypoints==0.4
exceptiongroup==1.1.0
filelock==3.9.0
fonttools==4.38.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.30
h11==0.14.0
huggingface-hub==0.12.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
Jinja2==3.1.2
JPype1==1.4.1
jsonschema==4.17.3
kiwisolver==1.4.4
konlpy==0.6.0
lxml==4.9.2
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.7.0
mdurl==0.1.2
multidict==6.0.4
numpy==1.24.2
openai==0.26.5
outcome==1.2.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==3.20.3
pyarrow==11.0.0
pycparser==2.21
pydeck==0.8.0
pydub==0.25.1
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
pytrends==4.9.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rich==13.3.1
scipy==1.10.0
selenium==4.8.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
streamlit==1.18.1
streamlit-drawable-canvas==0.9.2
streamlit-option-menu==0.3.2
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
torch==1.13.1
torchaudio==0.13.1
tornado==6.2
tqdm==4.64.1
transformers==4.26.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
watchdog==2.2.1
wincertstore==0.2
wordcloud==1.8.2.2
wsproto==1.2.0
yarl==1.8.2
zipp==3.13.0
